encapsulate player script

+++ Assets/Scripts/Scriptable Objects/SOEnemy.cs
@@ -6,7 +6,7 @@ public class SOEnemy : ScriptableObject
 {
     public int BaseEnemyHealth = 1;
     public int BaseEnemyDamage = 1;
-    public float BaseEnemyDetectionRadius = 50;
+    public int BaseEnemyDetectionRadius = 50;
     public float BaseEnemyAttackCooldown = 2F;
     public float BaseEnemySpeed = 3F;
